Given up on #34, but figured out ESLint instead & fixed small bug. Fixes #34
