+++ .gitignore
@@ -1,3 +1,4 @@
 /.vs
 /.vscode
 /.git
+/node_modules
