+++ LICENSE
@@ -1,6 +1,6 @@
 # MIT License
 
-## Copyright (c) 2023 mOrsExtension
+## Copyright (c) 2022-2024 mOrsExtension
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
