+++ background/helperBG.js
@@ -1,5 +1,4 @@
 //background/helperBG.js
-//@ts-check
 
 /** Sends infoTxt from background script to service worker inspection console; helperBG.js
  * @param {string} infoTxt
@@ -7,7 +6,7 @@
  * @param {string} calledBy
  * @param {string} color */
 const infoBG = (infoTxt, script = 'helperBG.js', calledBy = '', color = 'yellow') => {
-   console.info(`${script} ${calledBy}: %c${infoTxt}`, `color:${color}`)
+    console.info(`${script} ${calledBy}: %c${infoTxt}`, `color:${color}`)
 }
 
 /** Sends warnMsg information from background script to service worker inspection console; helperBG.js
@@ -16,23 +15,23 @@ const infoBG = (infoTxt, script = 'helperBG.js', calledBy = '', color = 'yellow'
  * @param {string} calledBy
  * @param {string} color */
 const warnBG = (warnTxt, script = 'helperBG.js', calledBy = '', color = '#9ff') => {
-   console.warn(`%c${script} ${calledBy}:%c ${warnTxt}`, 'color:#E46962', `color:${color}`)
+    console.warn(`%c${script} ${calledBy}:%c ${warnTxt}`, 'color:#E46962', `color:${color}`)
 }
 
 /**Listens for first installation; Deletes if necessary & reinstalls initial variables on new installation */
 browser.runtime.onInstalled.addListener(async details => {
-   try {
-      if (details.reason == 'install') {
-         await browser.storage.sync.clear()
-         const userInitData = await promiseReadJsonFile('userInitialData.json')
-         browser.storage.sync.set(userInitData) // no need to await this, nothing returned
-         infoBG('Set initial user variables from data/userInitialData.json')
-      }
-   } catch (error) {
-      warnBG(
-         `Failed to set user initial variables from data/userInitialData.json, ${error}`,
-         'helperBG.js',
-         'onInstalled.addListener'
-      )
-   }
+    try {
+        if (details.reason == 'install') {
+            await browser.storage.sync.clear()
+            const userInitData = await promiseReadJsonFile('userInitialData.json')
+            browser.storage.sync.set(userInitData) // no need to await this, nothing returned
+            infoBG('Set initial user variables from data/userInitialData.json')
+        }
+    } catch (error) {
+        warnBG(
+            `Failed to set user initial variables from data/userInitialData.json, ${error}`,
+            'helperBG.js',
+            'onInstalled.addListener'
+        )
+    }
 })
