+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "mors",
+  "version": "0.9.1",
+  "description": "Improves appearance of Oregon Revised Statutes on legislature's website and helps retrieve Oregon session laws.",
+  "main": "backgroundLoader.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "github.com/mOrsExtension/mOrs"
+  },
+  "keywords": [
+    "oregon",
+    "statutes",
+    "law",
+    "legislation"
+  ],
+  "author": "mOrsExtension@gmail.com",
+  "license": "MIT",
+  "devDependencies": {
+    "@eslint/js": "^9.2.0",
+    "eslint": "^9.2.0",
+    "globals": "^15.1.0"
+  }
+}
